Revise serializers to include all context
