+++ mig3/projects/tests/test_serializers.py
@@ -1,12 +1,13 @@
 from unittest import mock
 
+import projects.api.serializers.common
 from ..api import serializers
 
 
 def test_project_summary_serializer(project):
     """Should produce summary serialized Project on the happy path."""
     context = {"request": mock.Mock(name="request", GET={})}
-    serializer = serializers.ProjectSummarySerializer(instance=project, context=context)
+    serializer = projects.api.serializers.common.ProjectSummarySerializer(instance=project, context=context)
     assert serializer.data is not None
 
 
@@ -20,7 +21,7 @@ def test_project_serializer(project):
 def test_target_summary_serializer(target):
     """Should produce summary serialized Target on the happy path."""
     context = {"request": mock.Mock(name="request", GET={})}
-    serializer = serializers.TargetSummarySerializer(instance=target, context=context)
+    serializer = projects.api.serializers.common.TargetSummarySerializer(instance=target, context=context)
     assert serializer.data is not None
 
 
@@ -41,6 +42,6 @@ def test_version_read_serializer(version):
 def test_version_write_serializer(db):
     """Should produce Build with incoming serialized version request data."""
     data = {"hash": "a2" * 20, "author": "user@example.com"}
-    serializer = serializers.VersionWriteSerializer(data=data)
+    serializer = projects.api.serializers.common.VersionWriteSerializer(data=data)
     assert serializer.is_valid(raise_exception=True)
     assert serializer.save()
