+++ mig3/templates/index.html
@@ -15,8 +15,8 @@
     <noscript>
       <strong
         >We're sorry but frontend doesn't work properly without JavaScript
-        enabled. Please enable it to continue.</strong
-      >
+        enabled. Please enable it to continue.
+      </strong>
     </noscript>
     <div id="app">
       <app></app>
